晶格增加有机分子识别
